[zeroD] Move newFlowDevice to ConnectorFactory
