+++ include/cantera/zeroD/ConnectorFactory.h
@@ -12,6 +12,8 @@
 namespace Cantera
 {
 
+class FlowDevice;
+
 //! Factory class to create Connector objects.
 //!
 //! This class is mainly used via the newConnector() function, for example:
@@ -61,6 +63,15 @@ shared_ptr<Connector> newConnector(const string& model,
                                    shared_ptr<ReactorBase> r0,
                                    shared_ptr<ReactorBase> r1,
                                    const string& name="(none)");
+
+//! Create a FlowDevice object of the specified type
+//! @since Starting in %Cantera 3.1, this method returns a `shared_ptr<FlowDevice>`
+shared_ptr<FlowDevice> newFlowDevice(const string& model, const string& name="(none)");
+
+//! Create a FlowDevice object of the specified type
+//! @since New in %Cantera 3.0.
+//! @deprecated Replaced by newFlowDevice. To be removed after %Cantera 3.1.
+shared_ptr<FlowDevice> newFlowDevice3(const string& model);
 
 //! @}
 }
