+++ include/cantera/zerodim.h
@@ -33,7 +33,7 @@
 
 // factories
 #include "cantera/zeroD/ReactorFactory.h"
-#include "cantera/zeroD/FlowDeviceFactory.h"
+#include "cantera/zeroD/ConnectorFactory.h"
 #include "cantera/zeroD/WallFactory.h"
 
 // func1
