+++ src/zeroD/ConnectorFactory.cpp
@@ -4,6 +4,7 @@
 // at https://cantera.org/license.txt for license and copyright information.
 
 #include "cantera/zeroD/ConnectorFactory.h"
+#include "cantera/zeroD/FlowDevice.h"
 #include "cantera/zeroD/flowControllers.h"
 #include "cantera/zeroD/Wall.h"
 
@@ -50,5 +51,19 @@ shared_ptr<Connector> newConnector(
     return shared_ptr<Connector>(
         ConnectorFactory::factory()->create(model, r0, r1, name));
 }
+
+shared_ptr<FlowDevice> newFlowDevice(const string& model, const string& name)
+{
+    return std::dynamic_pointer_cast<FlowDevice>(
+        newConnector(model, nullptr, nullptr, name));
+}
+
+shared_ptr<FlowDevice> newFlowDevice3(const string& model)
+{
+    warn_deprecated("newFlowDevice3",
+        "Use newFlowDevice instead; to be removed after Cantera 3.1.");
+    return newFlowDevice(model);
+}
+
 
 }
